fix: resolve remaining go vet errors after stellar sub-package extraction
Move stellar_utils_test.go to stellar/ sub-package (functions moved in #18089),
add missing auth import in security_test.go, fix cherry-pick conflicts in
benchmarks_runner_test.go and mock_store.go.
