+++ pkg/api/handlers/stellar/stellar_utils_test.go
@@ -1,4 +1,4 @@
-package handlers
+package stellar
 
 import (
 	"bufio"
